+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/builtin/Builtins.java
@@ -45,7 +45,7 @@ import org.enso.pkg.QualifiedName;
 /** Container class for static predefined atoms, methods, and their containing scope. */
 public class Builtins {
   public static final String SOURCE_NAME = "Builtins.enso";
-  public static final String MODULE_NAME = "Builtins.Main";
+  public static final String MODULE_NAME = "Standard.Builtins.Main";
 
   /** Container for method names needed outside this class. */
   public static class MethodNames {
@@ -82,7 +82,7 @@ public class Builtins {
    */
   public Builtins(Context context) {
     Language language = context.getLanguage();
-    module = Module.empty(QualifiedName.fromString(MODULE_NAME));
+    module = Module.empty(QualifiedName.fromString(MODULE_NAME), null);
     scope = module.compileScope(context);
 
     any = new AtomConstructor("Any", scope).initializeFields();