+++ engine/runtime/src/test/scala/org/enso/interpreter/test/instrument/ReplTest.scala
@@ -19,7 +19,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "initialize properly" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main = Debug.breakpoint
           |""".stripMargin
@@ -30,7 +30,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "be able to execute arbitrary code in the caller scope" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    x = 1
@@ -50,7 +50,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "return the last evaluated value back to normal execution flow" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    a = 5
@@ -68,7 +68,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "be able to define its local variables" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    x = 10
@@ -86,7 +86,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "not overwrite bindings" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    x = 10
@@ -103,7 +103,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "access and modify monadic state" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |run =
           |    State.put Number 10
@@ -123,7 +123,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "be able to list local variables in its scope" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    x = 10
@@ -149,7 +149,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "be able to list bindings it has created" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    x = 10
@@ -176,7 +176,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "allow to be nested" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    10 * Debug.breakpoint + 1
@@ -199,7 +199,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "behave well when nested" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    x = 1
@@ -226,7 +226,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "handle errors gracefully" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    Debug.breakpoint
@@ -246,7 +246,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "attach language stack traces to the exception" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    Debug.breakpoint
@@ -272,7 +272,7 @@ class ReplTest extends InterpreterTest with BeforeAndAfter with EitherValues {
     "not pollute bindings upon nested error" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |main =
           |    Debug.breakpoint