+++ engine/runtime/src/test/scala/org/enso/interpreter/test/semantic/PolyglotTest.scala
@@ -11,7 +11,7 @@ class PolyglotTest extends InterpreterTest {
 
     "allow calling methods on static objects" in {
       val code =
-        """from Builtins import all
+        """from Standard.Builtins import all
           |
           |main =
           |    class = Java.lookup_class "org.enso.example.TestClass"
@@ -24,7 +24,7 @@ class PolyglotTest extends InterpreterTest {
 
     "allow instantiating objects and calling methods on them" in {
       val code =
-        """from Builtins import all
+        """from Standard.Builtins import all
           |
           |main =
           |    class = Java.lookup_class "org.enso.example.TestClass"
@@ -36,7 +36,7 @@ class PolyglotTest extends InterpreterTest {
 
     "allow listing available members of an object" in {
       val code =
-        """from Builtins import all
+        """from Standard.Builtins import all
           |
           |main =
           |    class = Java.lookup_class "org.enso.example.TestClass"