+++ RELEASES.md
@@ -2,6 +2,14 @@
 
 ## Interpreter/Runtime
 
+- Implemented changes to the import and export syntax, requiring to provide the
+  project namespace, or use the new `project` keyword to import from the current
+  project ([#1806](https://github.com/enso-org/enso/pull/1806)).
+
+# Enso 0.2.12 (2021-06-24)
+
+## Interpreter/Runtime
+
 - Upgraded the underlying runtime to
   [GraalVM 21.1.0](https://github.com/graalvm/graalvm-ce-builds/releases/tag/vm-21.1.0)
   ([#1738](https://github.com/enso-org/enso/pull/1738)). This brings a raft of
@@ -87,6 +95,12 @@
   features, but they should not be relied on as they are not considered stable.
   Only the 3 latest nightly builds are kept, so the nightly versions become
   obsolete very quickly.
+
+## Known Issues
+
+- This is a beta release, so please see the
+  [issue tracker](https://github.com/enso-org/enso/issues?q=is%3Aissue+is%3Aopen+created%3A%3C2021-06-24)
+  for issues opened before the release date.
 
 # Enso 0.2.11 (2021-04-28)